WvBdbHash::count() should return size_t instead int, like WvList and WvDict.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@3199 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/wvbdbhash.h
@@ -159,7 +159,7 @@ public:
     bool exists(const K &key)
         { return WvBdbHashBase::exists(datumize<K>(key)); }
 
-    int count()
+    size_t count()
     {
 	int res = 0;
 	Iter i(*this);
